Got the Create page working and it is populating in the database!
