+++ models/Ad.php
@@ -1,13 +1,13 @@
-Ad.php
+
 
 <?php
 
-require_once __DIR__ "/Model.php";
-require_once __DIR__ "/User.php";
+require_once __DIR__ . "/Model.php";
+require_once __DIR__ . "/User.php";
 
 class Ads extends Model
 {
-	protected static table = 'ads'
+	protected static $table = 'ads';
 // connects User to ads posted
 public function user() {
 	return User::find($this->userId);
@@ -17,7 +17,7 @@ public function user() {
 // will return te values of items
 public static function findAllUserById($id) {
 	self::dbConnect();
-	$query = 'SELECT * FROM' . static::$table . 'WHERE userID = :userId ORDER BY id desc':
+	$query = 'SELECT * FROM' . static::$table . 'WHERE userID = :userId ORDER BY id desc';
 	
 	$stmt = self::$dbc->prepare($query);
 	$stmt = bindValue(':userId', $id, PDO::PARAM_INT);
@@ -27,7 +27,7 @@ public static function findAllUserById($id) {
 	$results = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
 	//setting attributes of items object and getting $results content
-	$item = null:
+	$item = null;
 
 	if($results) {
 		$item = new static;
@@ -46,8 +46,8 @@ public static function displayFeaturedAds($num = 4) {
 
 	$results = $stmt->FetchAll(PDO::FETCH_ASSOC);
 
-	$item =null:
-
+	$item =null;
+	
 	if($results) {
 		$item = new static;
 		$item->attributes = $results;
@@ -57,6 +57,23 @@ public static function displayFeaturedAds($num = 4) {
 
 }
 
+// protected function insert() {
+//         // TODO: call dbConnect to ensure we have a database connection
+//         self::dbConnect();
+
+//         $insert = "INSERT INTO ads (itemName, price, description, sellerName, username)
+//         VALUES(:itemName, :price, :description, :sellerName, :username);";
+//         $stmt = self::$dbc->prepare($insert);
+//         $stmt->bindValue(':itemName', $this->itemName, PDO::PARAM_STR);
+//         $stmt->bindValue(':price', $this->price, PDO::PARAM_STR);
+//         $stmt->bindValue(':description', $this->description, PDO::PARAM_STR);
+//         $stmt->bindValue(':sellerName', $this->sellerName, PDO::PARAM_STR);    
+//         $stmt->bindValue(':username', $this->username, PDO::PARAM_STR);    
+        
+//         $stmt->execute(); 
+//         $this->id = self::$dbc->lastInsertId();
+
+//     }
